+++ CarServiceApplication/src/main/java/com/example/carserviceapplication/model/Car.java
@@ -16,5 +16,5 @@ public class Car {
     private String plate;
     private String model;
     private String brand;
-    private BigDecimal odometer;
+    private Integer odometer;
 }
